requirements.txt
openai>=1.1.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pyppeteer>=1.0.2
